requirements.txt
fastapi
sqlalchemy
mysql-connector-python  
aiomysql
uvicorn
python-dotenv